Better structure for Main vs InProcess Deephaven API servers.
Useful for #1810
