+++ server/src/main/java/io/deephaven/server/runner/DeephavenApiServerComponent.java
@@ -1,7 +1,6 @@
 package io.deephaven.server.runner;
 
 import dagger.BindsInstance;
-import dagger.Component;
 
 import javax.inject.Named;
 import java.io.PrintStream;
@@ -11,17 +10,12 @@ public interface DeephavenApiServerComponent {
     DeephavenApiServer getServer();
 
     interface Builder<B extends Builder<B>> {
-        @BindsInstance
-        B withPort(@Named("http.port") int port);
 
         @BindsInstance
         B withSchedulerPoolSize(@Named("scheduler.poolSize") int numThreads);
 
         @BindsInstance
         B withSessionTokenExpireTmMs(@Named("session.tokenExpireMs") long tokenExpireMs);
-
-        @BindsInstance
-        B withMaxInboundMessageSize(@Named("grpc.maxInboundMessageSize") int maxInboundMessageSize);
 
         @BindsInstance
         B withOut(@Named("out") PrintStream out);
