+++ server/src/main/java/io/deephaven/server/runner/InProcessModule.java
@@ -1,0 +1,10 @@
+package io.deephaven.server.runner;
+
+import dagger.Module;
+
+@Module(includes = {
+        DeephavenApiServerModule.class,
+        ServerBuilderInProcessModule.class
+})
+public interface InProcessModule {
+}
